+++ HISTORY.rst
@@ -1,8 +1,13 @@
 # History
 
-1.11.0 (2022-04-15)
+1.12.0 (2022-04-15)
 ------------------
 - BUGFIX: Import approximate_circle_by_area
+
+1.11.0 (2022-04-11)
+------------------
+- Add inner contours to labels2contours
+- Add inner contours to mask2json
 
 1.10.0 (2022-02-17)
 ------------------